+++ rosocp.go
@@ -1,13 +1,7 @@
 package main
 
-import (
-	"github.com/redhatinsights/ros-ocp-backend/cmd"
-	"github.com/redhatinsights/ros-ocp-backend/internal/config"
-	"github.com/redhatinsights/ros-ocp-backend/internal/plugin"
-)
+import "github.com/redhatinsights/ros-ocp-backend/cmd"
 
 func main() {
-	cfg := config.GetConfig()
-	plugin.ApplyLegacyUseNativeEngineEnv(cfg.UseNativeEngine)
 	cmd.Execute()
 }